Overload Random.float for 0, 1, and 2 arguments
Overloaded Random.float() in GlobalAPI.js so that it can handle being
called with 0, 1, or 2 arguments.

Random.float() with 0 arguments returns a random float in [0, 1).
Random.float(num1) with 1 argument returns a random float in the range
[0, num1).
Random.float(num1, num2) with 2 arguments returns a random float in the
range [num1, num2).

Additionally, a bug causing Random.float to generate numbers in the
range [num1, num2+1) was squashed.