Overload Random.integer for 0, 1, and 2 arguments
Overloaded Random.integer() in GlobalAPI.js so that it can handle being
called with 0, 1, or 2 arguments.

Random.integer() with 0 arguments randomly returns either 0 or 1.
Random.integer(num1) returns a random integer in the range [0, num1).
Random.integer(num1, num2) returns a random integer in the range [num1,
num2).

Additionally, the behavior of Random.integer was tweaked to be exclusive
of the maximum where before it was inclusive.