Fix zooming with zoom box
Zooming using the zoom box now works as expected. Drawing a box with the
zoom box causes the screen to zoom to where the edges of the box would
be.

Zooming is still a little confusing when zoomed in because there is a
zoom limit, and the zoom box does not zoom past the zoom limit. I think
it makes sense to keep the zoom limit, it's just something to be aware
of.

Fixes issue #59.